+++ pkg/sass-parser/CHANGELOG.md
@@ -12,6 +12,8 @@
 
 * Add support for parsing the `supports()` function in `@import` modifiers.
 
+* Add support for parsing unary operation expressions.
+
 ## 0.4.15
 
 * Add support for parsing list expressions.